requirements.txt
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
Pillow==8.4.0
numpy==1.21.2
